// eshop-libs/projects/sag-common/src/enums/keyboard.enum.ts
export enum KEY_BOARD {
    ENTER = 13,
    ARROW_DOWN = 40,
    ARROW_UP = 38,
    ZERO = 48,
    NUMBER_NINE = 57,
    SEPARATOR_COMMAS = 44,
    SEPARATOR_DOT = 46,
    BACK_SPACE = 8,
    ARROW_LEFT = 37,
    ARROW_RIGHT = 39,
    HOME = 36,
    END = 35,
    A = 65,
    Z = 90,
    a = 97,
    z = 122,
    SPACE = 32
}